backend/requirements.txt
asgiref==3.8.1
Django==4.2.10 
djangorestframework==3.16.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
python-dotenv ==1.1.0
django-autoslug==1.9.9
python-jose==3.3.0
django-cors-headers== 4.7.0
drf-yasg==1.21.7
setuptools==80.9.0